// plugins/plugin-chart-echarts/src/BigNumber/PieChartWithCards/types.ts
/**
 * Licensed to the Apache Software Foundation (ASF) under one
 * or more contributor license agreements.  See the NOTICE file
 * distributed with this work for additional information
 * regarding copyright ownership.  The ASF licenses this file
 * to you under the Apache License, Version 2.0 (the
 * "License"); you may not use this file except in compliance
 * with the License.  You may obtain a copy of the License at
 *
 *   http://www.apache.org/licenses/LICENSE-2.0
 *
 * Unless required by applicable law or agreed to in writing,
 * software distributed under the License is distributed on an
 * "AS IS" BASIS, WITHOUT WARRANTIES OR CONDITIONS OF ANY
 * KIND, either express or implied.  See the License for the
 * specific language governing permissions and limitations
 * under the License.
 */

import {
  QueryFormData,
  QueryFormMetric,
} from '@superset-ui/core';
import { BaseChartProps } from '../../types';

export interface PieChartDataItem {
  name: string;
  value: number;
}

export type PieChartWithCardsFormData = QueryFormData & {
  metric?: QueryFormMetric;
  title?: string;
  data?: PieChartDataItem[];
};

export type PieChartWithCardsChartProps =
  BaseChartProps<PieChartWithCardsFormData> & {
    formData: PieChartWithCardsFormData;
  };

export type PieChartWithCardsProps = {
  className?: string;
  width: number;
  height: number;
  data: PieChartDataItem[];
  title?: string;
};